Add ANTLR parser for comparison, not yet used
There is a strong possibility we will move to Antlr at some stage, to replace
reap, which is a place-holder while we are experimenting with GraphQL. If ANTLR
proved to provide a strong performance gain on top of other advantages (more
widespread developer knowledge), we'd probably switch sooner. However, these
performance benchmarks indicate that while Antlr is faster at raw parsing than
reap, as soon as you add post-process into Clojure-friendly
data-structures (something that the single-pass reap performs on-the-fly), then
rap is faster.